Handle case of broken directory symlinking
Symlinking of directories was broken. The problem was that the
originating file's basename was not appended to the destination
directory. As the destination is already a directory, the symlink
would fail.

So, update the code to opt out early for any symlinked directories
and let files continue on their origina path.